pyproject.toml
[project]
name = "2025-fastapi-seminar-template-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.116.1",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]
